Clean up some type definitions
