+++ backend/src/types/params.ts
@@ -10,3 +10,23 @@ export interface BranchSwitchArgs {
   url: string
   branch: string
 }
+
+export interface ServiceUserInput {
+  user_id: number
+  services_id: number
+  username: string
+  email: string
+  token: string
+  reposurl: string
+}
+
+export interface RegisterUserInput {
+  username: string
+  email: string
+  password: string
+}
+
+export interface LoginUserInput {
+  username: string
+  password: string
+}
